+++ mbed-client-pal/Source/Port/Reference-Impl/OS_Specific/SXOS/Storage/FileSystem/pal_plat_fileSystem.c
@@ -247,7 +247,7 @@ palStatus_t pal_plat_fsFwrite(palFileDescriptor_t *fd, const void *buffer, size_
 }
 
 
-palStatus_t pal_plat_fsFseek(palFileDescriptor_t *fd, int32_t offset, pal_fsOffset_t whence)
+palStatus_t pal_plat_fsFseek(palFileDescriptor_t *fd, off_t offset, pal_fsOffset_t whence)
 {
     palStatus_t status = PAL_SUCCESS;
 
@@ -297,7 +297,7 @@ palStatus_t pal_plat_fsFseek(palFileDescriptor_t *fd, int32_t offset, pal_fsOffs
 }
 
 
-palStatus_t pal_plat_fsFtell(palFileDescriptor_t *fd, int32_t *pos)
+palStatus_t pal_plat_fsFtell(palFileDescriptor_t *fd, off_t *pos)
 {
     palStatus_t status = PAL_SUCCESS;
